Chore: move struct documentation to own package (#17)
* chore: move struct doc to own package

---------
